requirements.txt
# Streamlit関連
streamlit>=1.28.0
streamlit-webrtc>=0.44.0

# 映像処理
opencv-python-headless>=4.8.0
deepface>=0.0.79

# 音声処理・AI
openai>=1.3.0
# librosa>=0.10.0  # オプション（トーン分析用、後回し可）

# その他
python-dotenv>=1.0.0
numpy>=1.24.0
av>=10.0.0

